+++ mathesar_ui/src/pages/record/RecordViewLoadingSpinner.svelte
